Tell api when we are done using a token
refs: none
affects: Parent, Teacher, Student
release note: Fixed an issue causing orphaned Approved Integrations for Canvas for IOS
