+++ Core/Core/API/APIOAuth.swift
@@ -41,3 +41,112 @@ public struct APIOAuthUser: Codable, Equatable {
     let effective_locale: String
     let email: String?
 }
+
+// Not documented in canvas rest api
+struct GetMobileVerifyRequest: APIRequestable {
+    typealias Response = APIVerifyClient
+
+    let domain: String
+
+    let path = "https://canvas.instructure.com/api/v1/mobile_verify.json"
+    var query: [APIQueryItem] {
+        return [.value("domain", domain)]
+    }
+    var cachePolicy: URLRequest.CachePolicy {
+        return .reloadIgnoringLocalAndRemoteCacheData
+    }
+    let headers: [String: String?] = [
+        HttpHeader.accept: "application/json",
+        HttpHeader.authorization: nil,
+    ]
+}
+
+// https://canvas.instructure.com/doc/api/file.oauth_endpoints.html#post-login-oauth2-token
+struct PostLoginOAuthRequest: APIRequestable {
+    typealias Response = APIOAuthToken
+
+    enum GrantType {
+        case code(String)
+        case refreshToken(String)
+    }
+
+    let client: APIVerifyClient
+    let grantType: GrantType
+
+    init(client: APIVerifyClient, code: String) {
+        self.client = client
+        self.grantType = .code(code)
+    }
+
+    init(client: APIVerifyClient, refreshToken: String) {
+        self.client = client
+        self.grantType = .refreshToken(refreshToken)
+    }
+
+    let method = APIMethod.post
+    var path: String {
+        return URL(string: "login/oauth2/token", relativeTo: client.base_url)?.absoluteString ?? "login/oauth2/token"
+    }
+    var query: [APIQueryItem] {
+        var query: [APIQueryItem] = [
+            .value("client_id", client.client_id ?? ""),
+            .value("client_secret", client.client_secret ?? ""),
+        ]
+
+        switch grantType {
+        case .code(let code):
+            query.append(.value("grant_type", "authorization_code"))
+            query.append(.value("code", code))
+        case .refreshToken(let token):
+            query.append(.value("grant_type", "refresh_token"))
+            query.append(.value("refresh_token", token))
+        }
+
+        return query
+    }
+    let headers: [String: String?] = [
+        HttpHeader.authorization: nil,
+    ]
+}
+
+// https://canvas.instructure.com/doc/api/file.oauth_endpoints.html#delete-login-oauth2-token
+public struct DeleteLoginOAuthRequest: APIRequestable {
+    public typealias Response = APINoContent
+
+    let session: LoginSession
+    public init(session: LoginSession) {
+        self.session = session
+    }
+
+    public let method = APIMethod.delete
+    public var path: String {
+        return session.baseURL.appendingPathComponent("login/oauth2/token").absoluteString
+    }
+
+    public var headers: [String: String?] {
+        return [
+            HttpHeader.authorization: session.accessToken.flatMap { "Bearer \($0)" },
+        ]
+    }
+}
+
+// https://canvas.instructure.com/doc/api/file.oauth_endpoints.html#get-login-session-token
+public struct GetWebSessionRequest: APIRequestable {
+    public struct Response: Codable {
+        public let session_url: URL
+    }
+
+    public let to: URL?
+
+    public let path = "/login/session_token"
+    public var query: [APIQueryItem] {
+        if let returnTo = to?.appendingQueryItems(URLQueryItem(name: "display", value: "borderless")) {
+            return [ .value("return_to", returnTo.absoluteString) ]
+        }
+        return []
+    }
+
+    public init(to: URL?) {
+        self.to = to
+    }
+}
