+++ Student/Student/CanvasAppDelegate.swift
@@ -383,6 +383,7 @@ extension AppDelegate: LoginDelegate, NativeLoginManagerDelegate {
 
     func userDidStopActing(as session: LoginSession) {
         LoginSession.remove(session)
+        environment.api.makeRequest(DeleteLoginOAuthRequest(session: session)) { _, _, _ in }
         guard environment.currentSession == session else { return }
         PageViewEventController.instance.userDidChange()
         NotificationKitController.deregisterPushNotifications { _ in }
